Refactor to make builds optimized
