+++ pages/_app.tsx
@@ -1,7 +1,7 @@
 import '../styles/globals.css'
 import type {AppProps} from 'next/app'
 import {createTheme, ThemeProvider} from "@mui/material";
-import {AppProvider} from "./util/app-context";
+import {AppProvider} from "../lib/util/app-context";
 
 const darkTheme = createTheme({
     palette: {
